chore: removes dynamic requires
