+++ packages/react/src/__tests__/server-side-hydrate.test.tsx
@@ -1,31 +1,21 @@
 import React from 'react';
-import { renderToString } from 'react-dom/server';
-import ReactDOM from 'react-dom';
-import { styled } from '@compiled/react';
-import { isNodeEnvironment } from '@compiled/react/dist/runtime/is-node';
-import { useCache } from '@compiled/react/dist/runtime/provider';
+import { hydrate } from 'react-dom';
+import { CC, CS } from '../runtime';
 
-jest.mock('@compiled/react/dist/runtime/is-node');
+jest.mock('../runtime/is-node', () => ({
+  isNodeEnvironment: () => false,
+}));
 
 describe('server side hydrate', () => {
   beforeEach(() => {
     jest.spyOn(global.console, 'error');
-    flushEnvironment('node');
   });
 
-  const flushEnvironment = (env: 'node' | 'browser') => {
-    // This isn't a real hook.
-    // eslint-disable-next-line react-hooks/rules-of-hooks
-    const cache = useCache();
-    for (const key in cache) {
-      // Flush the cache out - unfortunately it persisted between tests.
-      delete cache[key];
-    }
-    (isNodeEnvironment as jest.Mock).mockReturnValue(env === 'node');
+  const flushRuntimeModule = () => {
     jest.resetModules();
     // We need to force this module to re-instantiate because on the client
     // when it does it will move all found SSRd style elements to the head.
-    require('@compiled/react/runtime');
+    require('../runtime');
   };
 
   const appendHTML = (markup: string) => {
@@ -36,16 +26,22 @@ describe('server side hydrate', () => {
   };
 
   it('should not log any warnings when hydrating HTML', () => {
-    const StyledDiv = styled.div`
-      font-size: 12px;
-      color: blue;
-      border: 1px solid black;
-    `;
+    // It's notoriously hard to do both server and client rendering in this test.
+    // Instead of doing the server flow we hardcode the result instead.
+    const elem = appendHTML(
+      `<style data-cmpld="true">.foo{color:blue}</style><div class="foo">hello world</div>`
+    );
 
-    const element = <StyledDiv>hello world</StyledDiv>;
-    const app = appendHTML(renderToString(element));
-    flushEnvironment('browser');
-    ReactDOM.hydrate(element, app);
+    flushRuntimeModule();
+    hydrate(
+      <CC>
+        <CS>{['.foo { color: blue; }', '.foo { color: blue; }']}</CS>
+        <div className="foo">hello world</div>
+      </CC>,
+      elem
+    );
+
+    console.log(document.body.innerHTML);
 
     expect(console.error).not.toHaveBeenCalled();
   });
