+++ packages/react/src/runtime/index.tsx
@@ -1,3 +1,3 @@
 export { default as CS } from './style';
-export { default as CC } from './provider';
+export { default as CC } from './style-cache';
 export { default as ax } from './ax';
