--- packages/react/src/runtime/provider/types.tsx
@@ -1,3 +1,0 @@
-export type UseCacheHook = () => Record<string, true>;
-
-export type ProviderComponent = (props: { children: JSX.Element[] | JSX.Element }) => JSX.Element;
